+++ src/Views/Form/Templates/Sequoia/optionConfig.php
@@ -104,6 +104,17 @@ return [
 				],
 				'default'    => __( 'Continue', 'give' ),
 			],
+			[
+				'id'      => 'decimals_enabled',
+				'name'    => __( 'Decimal amounts', 'give' ),
+				'desc'    => __( 'Do you want to enable decimal amounts? When the setting is disabled, decimal values are rounded.', 'give' ),
+				'type'    => 'radio_inline',
+				'default' => 'disabled',
+				'options' => [
+					'disabled' => __( 'Disabled', 'give' ),
+					'enabled'  => __( 'Enabled', 'give' ),
+				],
+			],
 		],
 	],
 	'payment_information' => [